Refs #106. set MCVINE_MPI_BINDING to NONE when MCVINE_MPI_LAUNCHER is set to "serial". also updated tests/mcstas2/pyre_support/componentfactory_TestCase.py
